fix rename leading to error
