+++ src/utility/createMediaUrlObjectFromFilePath.ts
